feat(onboarding): channel routing in slack step with progressive reveal
The slack step now lets users route notifications without leaving
onboarding. `incidents` (the group the step is selling) shows as
soon as Slack is connected; `wrapped_reports` and `custom_messages`
appear once any route has been set, so users aren't faced with three
unfamiliar rows on first sight.

Extracted `SlackRouteRow` out of the settings page into a shared
component file so onboarding and settings stay visually identical
and a save in either place invalidates the same `["slack-integration"]`
query key. Disconnect remains a settings-only affordance — onboarding
isn't the place to undo a connection that was just made. Footer note
points users to settings for the full picture.
